static loop 수정
